Update util_json.py
This added the support to generate JSON data based on function defintion
and class defintion in Python files. The API is transform(path), in which
path is a Python file or a dir (in later case all Python files in the dir
and its subdirs are processed). Json files created by the function are
located in the same diretory as their corresponding Python source files.
The function outputs json file names on stdout on success and error log
on stderr if it fails to parse Python source files.

Below is a simple example on how to use the API to write a command line tool:

    import sys
    if len(sys.argv) != 2:
        print("Usage: %s <path>" % sys.argv[0])
        sys.exit(0)
    transform(sys.argv[1])